+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- `do` block function
+
 ## v1.0.0
 
 - `forks` block function [10ab025](https://github.com/gabiseabra/redux-saga-test-factory/commit/10ab025)